+++ test/fixtures/cookbooks/test/recipes/openjdk.rb
@@ -1,1 +1,3 @@
 openjdk_install node['version']
+
+include_recipe 'test::java_cert'